@@ -20,35 +20,34 @@
   },
   "homepage": "https://github.com/woutervh-/jimney-web#readme",
   "devDependencies": {
-    "@types/react": "^16.7.2",
-    "@types/react-dom": "^16.0.9",
-    "@types/react-router-dom": "^4.3.1",
+    "@types/react": "^17.0.45",
+    "@types/react-dom": "^17.0.16",
     "css-loader": "^0.28.11",
     "file-loader": "^1.1.11",
-    "html-webpack-plugin": "^3.2.0",
+    "html-webpack-plugin": "^5.5.0",
     "jimp": "^0.2.28",
     "jsdom": "^11.12.0",
     "mini-css-extract-plugin": "^0.4.4",
-    "node-sass": "^4.10.0",
-    "postcss-cssnext": "^3.1.0",
-    "postcss-loader": "^2.1.6",
-    "react": "^16.6.1",
-    "react-dom": "^16.6.1",
-    "react-router-dom": "^4.3.1",
+    "node-sass": "^7.0.1",
+    "postcss-cssnext": "^3.1.1",
+    "postcss-loader": "^6.2.1",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "react-router-dom": "^6.3.0",
     "responsive-loader": "^1.1.0",
     "rimraf": "^2.6.2",
-    "sass-loader": "^7.1.0",
-    "style-loader": "^0.21.0",
+    "sass-loader": "^12.6.0",
+    "style-loader": "^3.3.1",
     "sugarss": "^1.0.1",
-    "svgo": "^1.1.1",
-    "svgo-loader": "^2.2.0",
-    "ts-loader": "^4.5.0",
-    "typescript": "^2.9.2",
+    "svgo": "^2.8.0",
+    "svgo-loader": "^3.0.0",
+    "ts-loader": "^9.3.0",
+    "typescript": "^4.6.4",
     "uglifyjs-webpack-plugin": "^1.3.0",
-    "url-loader": "^1.1.2",
-    "webpack": "^4.25.1",
-    "webpack-cli": "^3.1.2",
-    "webpack-dev-server": "^3.1.10",
-    "webpack-merge": "^4.1.4"
+    "url-loader": "^4.1.1",
+    "webpack": "^5.72.0",
+    "webpack-cli": "^4.9.2",
+    "webpack-dev-server": "^4.9.0",
+    "webpack-merge": "^5.8.0"
   }
 }
