@@ -1,5 +1,5 @@
 const path = require('path');
-const merge = require('webpack-merge');
+const { merge } = require('webpack-merge');
 const webpack = require('webpack');
 const common = require('./webpack.common.js');
 const MiniCssExtractPlugin = require('mini-css-extract-plugin');
@@ -38,7 +38,6 @@ module.exports = merge(common, {
         new webpack.DefinePlugin({
             PRODUCTION: false
         }),
-        new webpack.HotModuleReplacementPlugin(),
         new HtmlWebpackPlugin({ template: './src/index.html', chunks: ['bundle'], filename: 'index.html' }),
         new MiniCssExtractPlugin({
             filename: '[name].css',
